refactor: move computeName to build
